+++ Source/Swig/typeobj.c
@@ -361,7 +361,7 @@ SwigType *SwigType_del_pointer(SwigType *t) {
     printf("Fatal error. SwigType_del_pointer applied to non-pointer.\n");
     abort();
   }
-  Delslice(t, 0, (c - s) + 2);
+  Delslice(t, 0, (int)((c - s) + 2));
   return t;
 }
 
@@ -915,7 +915,7 @@ SwigType *SwigType_add_template(SwigType *t, ParmList *parms) {
 String *SwigType_templateprefix(const SwigType *t) {
   const char *s = Char(t);
   const char *c = strstr(s, "<(");
-  return c ? NewStringWithSize(s, c - s) : NewString(s);
+  return c ? NewStringWithSize(s, (int)(c - s)) : NewString(s);
 }
 
 /* -----------------------------------------------------------------------------
@@ -966,7 +966,7 @@ String *SwigType_templatesuffix(const SwigType *t) {
 String *SwigType_istemplate_templateprefix(const SwigType *t) {
   const char *s = Char(t);
   const char *c = strstr(s, "<(");
-  return c ? NewStringWithSize(s, c - s) : 0;
+  return c ? NewStringWithSize(s, (int)(c - s)) : 0;
 }
 
 /* -----------------------------------------------------------------------------
@@ -989,7 +989,7 @@ String *SwigType_istemplate_only_templateprefix(const SwigType *t) {
   const char *s = Char(t);
   if (len >= 4 && strcmp(s + len - 2, ")>") == 0) {
     const char *c = strstr(s, "<(");
-    return c ? NewStringWithSize(s, c - s) : 0;
+    return c ? NewStringWithSize(s, (int)(c - s)) : 0;
   } else {
     return 0;
   }
@@ -1022,7 +1022,7 @@ String *SwigType_templateargs(const SwigType *t) {
 	  nest--;
 	c++;
       }
-      return NewStringWithSize(start, c - start);
+      return NewStringWithSize(start, (int)(c - start));
     }
     c++;
   }
